+++ spec/check_localization_progress_spec.rb
@@ -9,9 +9,9 @@ describe Fastlane::Actions::CheckTranslationProgressAction do
 
   it 'does not fail when all the languages are above the set threshold' do
     langs = [
-      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0' },
-      { lang_name: 'german', lang_code: 'de', current: '2,078', fuzzy: '2', waiting: '3', untranslated: '4' },
-      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2' },
+      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0', progress: '100' },
+      { lang_name: 'german', lang_code: 'de', current: '2,078', fuzzy: '2', waiting: '3', untranslated: '4', progress: '99' },
+      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2', progress: '99' },
     ]
 
     stub = stub_request(
@@ -38,8 +38,8 @@ describe Fastlane::Actions::CheckTranslationProgressAction do
 
   it 'fails on missing data for a language' do
     langs = [
-      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0' },
-      { lang_name: 'german', lang_code: 'de', current: '2,078', fuzzy: '2', waiting: '3', untranslated: '4' },
+      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0', progress: '100' },
+      { lang_name: 'german', lang_code: 'de', current: '2,078', fuzzy: '2', waiting: '3', untranslated: '4', progress: '99' },
     ]
 
     stub = stub_request(
@@ -95,10 +95,9 @@ describe Fastlane::Actions::CheckTranslationProgressAction do
 
   it 'fails when one the language is below the set threshold' do
     langs = [
-      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0' },
-      # Mock de to be translated at 51%
-      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4' },
-      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2' },
+      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0', progress: '100' },
+      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4', progress: '51' },
+      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2', progress: '99' },
     ]
 
     stub = stub_request(
@@ -123,12 +122,40 @@ describe Fastlane::Actions::CheckTranslationProgressAction do
     expect(stub).to have_been_made.once
   end
 
+  it 'does not fail when one the language is above the set threshold, but there are updates waiting for review' do
+    langs = [
+      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0', progress: '100' },
+      { lang_name: 'german', lang_code: 'de', current: '2,087', fuzzy: '0', waiting: '103', untranslated: '0', progress: '99' },
+      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2', progress: '99' },
+    ]
+
+    stub = stub_request(
+      :get,
+      'https://translate.wordpress.org/projects/apps/my-test-project/dev'
+    )
+           .to_return(
+             status: 200,
+             body: generate_glotpress_response_body(languages: langs)
+           )
+
+    expect(FastlaneCore::UI).to receive(:success).with('Done')
+
+    described_class.run(
+      glotpress_url: 'https://translate.wordpress.org/projects/apps/my-test-project/dev',
+      language_codes: 'ar de es'.split(),
+      min_acceptable_translation_percentage: 99,
+      abort_on_violations: true,
+      skip_confirm: true
+    )
+
+    expect(stub).to have_been_made.once
+  end
+
   it 'prints the report and asks user confirmation when one the language is below the threshold and aborting is disabled' do
     langs = [
-      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0' },
-      # Mock de to be translated at 51%
-      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4' },
-      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2' },
+      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0', progress: '100' },
+      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4', progress: '51' },
+      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2', progress: '99' },
     ]
 
     stub = stub_request(
@@ -162,11 +189,9 @@ describe Fastlane::Actions::CheckTranslationProgressAction do
 
   it 'prints the report and asks user confirmation when multiples languages are below the threshold and aborting is disabled' do
     langs = [
-      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0' },
-      # Mock de to be translated at 51%
-      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4' },
-      # Mock de to be translated at 75%
-      { lang_name: 'spanish', lang_code: 'es', current: '1,585', fuzzy: '0', waiting: '0', untranslated: '502' },
+      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0', progress: '100' },
+      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4', progress: '51' },
+      { lang_name: 'spanish', lang_code: 'es', current: '1,585', fuzzy: '0', waiting: '0', untranslated: '502', progress: '75' },
     ]
 
     stub = stub_request(
@@ -201,11 +226,9 @@ describe Fastlane::Actions::CheckTranslationProgressAction do
 
   it 'prints the report and continues when one the language is below the threshold, aborting is disabled and confirmation is skipped' do
     langs = [
-      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0' },
-      # Mock de to be translated at 51%
-      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4' },
-      # Mock de to be translated at 75%
-      { lang_name: 'spanish', lang_code: 'es', current: '1,585', fuzzy: '0', waiting: '0', untranslated: '502' },
+      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0', progress: '100' },
+      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4', progress: '51' },
+      { lang_name: 'spanish', lang_code: 'es', current: '1,585', fuzzy: '0', waiting: '0', untranslated: '502', progress: '75' },
     ]
 
     stub = stub_request(
@@ -239,10 +262,9 @@ describe Fastlane::Actions::CheckTranslationProgressAction do
 
   it 'prints the report and continues when multiple languages are below the threshold, aborting is disabled and confirmation is skipped' do
     langs = [
-      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0' },
-      # Mock de to be translated at 51%
-      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4' },
-      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2' },
+      { lang_name: 'arabic', lang_code: 'ar', current: '2,087', fuzzy: '0', waiting: '0', untranslated: '0', progress: '100' },
+      { lang_name: 'german', lang_code: 'de', current: '1,078', fuzzy: '2', waiting: '1003', untranslated: '4', progress: '51' },
+      { lang_name: 'spanish', lang_code: 'es', current: '2,085', fuzzy: '0', waiting: '0', untranslated: '2', progress: '100' },
     ]
 
     stub = stub_request(
@@ -284,7 +306,8 @@ def generate_glotpress_response_body(languages:)
       current: language[:current],
       fuzzy: language[:fuzzy],
       waiting: language[:waiting],
-      untranslated: language[:untranslated]
+      untranslated: language[:untranslated],
+      progress: language[:progress]
     )
   end
 
@@ -311,32 +334,29 @@ def generate_glotpress_response_header
   HEADER
 end
 
-def generate_glotpress_response_for_language(lang:, lang_code:, current:, fuzzy:, waiting:, untranslated:)
-  lang << <<~LANG
-    <tr class="odd">
-    		<td>
-  LANG
+def generate_glotpress_response_for_language(lang:, lang_code:, current:, fuzzy:, waiting:, untranslated:, progress:)
+  res = "<tr class=\"odd\">\n"
+  res << generate_glotpress_response_header_for_language(lang: lang, lang_code: lang_code, progress: progress)
+  res << generate_glotpress_response_for_language_status(lang_code: lang_code, status_main: 'translated', status: 'current', string_count: current)
+  res << generate_glotpress_response_for_language_status(lang_code: lang_code, status_main: 'fuzzy', status: 'fuzzy', string_count: fuzzy)
+  res << generate_glotpress_response_for_language_status(lang_code: lang_code, status_main: 'untranslated', status: 'untranslated', string_count: waiting)
+  res << generate_glotpress_response_for_language_status(lang_code: lang_code, status_main: 'waiting', status: 'waiting', string_count: untranslated)
+  res <<	'</tr>'
+end
 
-  lang << "<strong><a href=\"/projects/apps/android/dev/#{lang_code}/default/\">#{lang}</a></strong>\n"
-
-  lang << <<~LANG
-        <span class="bubble morethan90">anyperc%</span>
-    </td>
-    <td class="stats percent">anyperc%</td>
-  LANG
-
-  lang << generate_glotpress_response_for_language_status(lang_code: lang_code, status_main: 'translated', status: 'current', string_count: current)
-  lang << generate_glotpress_response_for_language_status(lang_code: lang_code, status_main: 'fuzzy', status: 'fuzzy', string_count: fuzzy)
-  lang << generate_glotpress_response_for_language_status(lang_code: lang_code, status_main: 'untranslated', status: 'untranslated', string_count: waiting)
-  lang << generate_glotpress_response_for_language_status(lang_code: lang_code, status_main: 'waiting', status: 'waiting', string_count: untranslated)
-  lang <<	'</tr>'
+def generate_glotpress_response_header_for_language(lang:, lang_code:, progress:)
+  res = "<td>\n"
+  res << "<strong><a href=\"/projects/apps/whatever/dev/#{lang_code}/default/\">#{lang}</a></strong>\n"
+  res << "<span class=\"bubble morethan90\">#{progress}%</span>\n" if progress.to_i > 90
+  res << "</td>\n"
+  res << "<td class=\"stats percent\">#{progress}%</td>\n"
 end
 
 def generate_glotpress_response_for_language_status(lang_code:, status_main:, status:, string_count:)
-  lang = "<td class=\"stats #{status_main}\" title=\"#{status_main}\">\n"
-  lang << "<a href=\"/projects/apps/whatever/dev/#{lang_code}/default/?filters%5Btranslated%5D=yes&#038;filters%5Bstatus%5D=#{status}\">#{string_count}</a></td>"
+  res = "<td class=\"stats #{status_main}\" title=\"#{status_main}\">\n"
+  res << "<a href=\"/projects/apps/whatever/dev/#{lang_code}/default/?filters%5Btranslated%5D=yes&#038;filters%5Bstatus%5D=#{status}\">#{string_count}</a></td>"
 
-  lang
+  res
 end
 
 def generate_glotpress_response_footer