+++ CHANGELOG.md
@@ -16,6 +16,7 @@ _None_
 
 * Fix crashes introduced in `1.3.0` – incorrect parameters in calls to `get_release_version`. [#283]
 * Fix the way versioning is handled for alphas – i.e. `version.properties` is indexed by flavor name, defaulting to `zalpha` for alphas. [#283]
+* Fixed an issue in `check_translation_progress` where a wrong evaluation of the progress is possible when there are Waiting string in GlotPress.
 
 ### Internal Changes